# server/instances/retrieve/src/api/services/User.js
/**
 * @param {Object} options
 * @param {String} options.id the &#x60;id&#x60; of the &#x60;User&#x60; to delete
 * @param {Function} callback
 */



import models from '../../../../../model';
import bcrypt from 'bcrypt';

const Grant = models.grant;
const Post = models.post;
const Project = models.project;
const Review = models.review;
const Skill = models.skill;
const University = models.University;
const User = models.User;
const UserSkill = models.userkSkill;




 
export function getUserById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getCityNameById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.city The city
 * @param {Function} callback
 */
export function isUserCity (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function isCollaboratorById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getCreatedDateById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function isCreatorById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getFirstNameById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.first_name The keywords
 * @param {Function} callback
 */
export function isUserFirstName (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getAllUserImages (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getAllUserKeywords (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.keyword The keyword
 * @param {Function} callback
 */
export function isUserHaveKeyword (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getLastNameById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.last_name The keywords
 * @param {Function} callback
 */
export function isUserLastName (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getModifiedDateById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getAllUserProjectById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.project_id The &#x60;project_id&#x60; of the &#x60;Project&#x60; to retrieve
 * @param {Function} callback
 */
export function isUserOnProjectById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getAllUserReviewById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.review_id The &#x60;review_id&#x60; of the &#x60;Project&#x60; to retrieve
 * @param {Function} callback
 */
export function isReviewOnUserById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getAllUserSkills (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.skill_id The &#x60;id&#x60; of the &#x60;Skill&#x60; to retrieve
 * @param {Function} callback
 */
export function getIsSkillOnUser (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {Function} callback
 */
export function getStateById (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.state The state
 * @param {Function} callback
 */
export function isUserState (options, callback) {
  // Implement you business logic here...
}

/**
 * @param {Object} options
 * @param {String} options.id The &#x60;id&#x60; of the &#x60;User&#x60; to retrieve
 * @param {String} options.university_id The &#x60;university_id&#x60; of the &#x60;University&#x60; to retrieve
 * @param {Function} callback
 */
export function isUniversityOnUser (options, callback) {
  // Implement you business logic here...
}