# swagger_ui/retrieve/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Definition of ArmRay backend API for data retrieval operations"
  version: "0.0.1"
  title: "ArmRay API"
host: "ec2-52-36-38-196.us-west-2.compute.amazonaws.com:1151"
basePath: "/armray/api/0.1/rest/retrieve"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /Project/{id}/reviews/{review_id}:
    get:
      tags:
      - "Project"
      - "Review"
      description: "Check if `Review` of a `Project` if it is on project\n"
      operationId: "isReviewOnOProject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `Project` to retrieve"
        required: true
        type: "string"
      - name: "review_id"
        in: "path"
        description: "The `id` of the review"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Project"
...
  /Review/{id}:
    get:
      tags:
      - "Review"
      description: "Retrieves `Review` objects by **id**.\n"
      operationId: "getReviewById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` name"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Review`"
          schema:
            $ref: "#/definitions/Review"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
  /Review/{id}/body:
    get:
      tags:
      - "Review"
      description: "Sets body of review\n"
      operationId: "getReviewBodyById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Review` body"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
...
  /Review/{id}/rating:
    get:
      tags:
      - "Review"
      description: "Gets rating of review\n"
      operationId: "getReviewRatingById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Review` rating"
          schema:
            type: "number"
            format: "int32"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
  /Review/{id}/rating/{rating}:
    get:
      tags:
      - "Review"
      description: "Is rating equal to what we pass in\n"
      operationId: "isReviewRating"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      - name: "rating"
        in: "path"
        description: "The `Review` rating"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
  /Review/{id}/reviewer:
    get:
      tags:
      - "Review"
      - "User"
      description: "Gets reviewer of review\n"
      operationId: "getReviewReviewerById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Review` reviewer `User`id"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
  /Review/{id}/reviewer/{user_id}:
    get:
      tags:
      - "Review"
      - "User"
      description: "Is `User`the reviewer?\n"
      operationId: "isUserReviewReviewer"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      - name: "user_id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
...
  /Review/{id}/url:
    get:
      tags:
      - "Review"
      description: "Gets url of review\n"
      operationId: "getReviewUrlById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Review` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Review` url"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Review"
...
  /User/{id}:
    get:
      tags:
      - "User"
      description: "Retrieve a 'User' by their id.\n"
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "the `id` of the `User` to delete"
        required: true
        type: "string"
      responses:
        200:
          description: "The `id` of the deleted `User`"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /User/{id}/created:
    get:
      tags:
      - "User"
      description: "Retrieve creation date of a `User`.\n"
      operationId: "getCreatedDateById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The date"
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /User/{id}/keys:
    get:
      tags:
      - "User"
      description: "Retrieve all keywords associated with a `User`.\n"
      operationId: "getAllUserKeywords"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of keywords"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /User/{id}/review/{review_id}:
    get:
      tags:
      - "User"
      - "Review"
      description: "Is a 'Review' associated with a `User`??\n"
      operationId: "isReviewOnUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "review_id"
        in: "path"
        description: "The `review_id` of the `Project` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The boolean"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
  /User/{id}/skills:
    get:
      tags:
      - "User"
      - "Skill"
      description: "Retrieve all skills associated with a `User`.\n"
      operationId: "getAllUserSkills"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of UserSkills ids"
          schema:
            type: "array"
            items:
              type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /User/{id}/uni/{university_id}:
    get:
      tags:
      - "User"
      - "University"
      description: "Is 'University' associated with a `User`?\n"
      operationId: "isUniversityOnUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `id` of the `User` to retrieve"
        required: true
        type: "string"
      - name: "university_id"
        in: "path"
        description: "The `university_id` of the `University` to retrieve"
        required: true
        type: "string"
      responses:
        200:
          description: "The `University` associated with the `User`"
          schema:
            type: "boolean"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "User"
...
  /UserSkill/{id}/user:
    get:
      tags:
      - "Skill"
      - "UserSkill"
      - "User"
      description: "Gets the `user`\n"
      operationId: "getUserSkillUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The `Skill` id"
        required: true
        type: "string"
      responses:
        200:
          description: "The `Skill`"
          schema:
            $ref: "#/definitions/Skill"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Skill"
...
definitions:
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      fields:
        type: "string"
      message:
        type: "string"
...
  Review:
    type: "object"
    properties:
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Review`creation"
      rating:
        type: "integer"
        format: "int32"
        description: "rating, must be 1-5"
      review_body:
        type: "string"
        description: "text body of review"
      review_id:
        type: "string"
        description: "id of review"
      reviewer:
        description: "reviewer user"
        $ref: "#/definitions/User"
      reviewer_id:
        type: "string"
        description: "reviewer `User`id"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Review`update"
...
  Skill:
    type: "object"
    properties:
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Skill`creation"
      description:
        type: "string"
        description: "Description of `Skill`"
      review_id:
        type: "string"
        description: "id of skill"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `Skill`update"
...
  User:
    type: "object"
    properties:
      city:
        type: "string"
        description: "City of `User`location"
      created_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `User`creation"
      first_name:
        type: "string"
        description: "User's first name."
      geolocation:
        type: "array"
        description: "Tuple representing latitude/longitude of user"
        items:
          type: "number"
          format: "double"
      is_collaborator:
        type: "boolean"
        description: "Is `User`a `Project`collaborator?"
      is_creator:
        type: "boolean"
        description: "Is `User`a `Project`creator?"
      is_user_private:
        type: "boolean"
        description: "is `User`profile public or private?"
      keywords:
        type: "array"
        description: "array of keywords associated with the user"
        items:
          type: "string"
      last_name:
        type: "string"
        description: "User's last name"
      profile_images:
        type: "array"
        description: "Array of file system paths to profile images"
        items:
          type: "string"
      project_ids:
        type: "array"
        description: "array of `User`projects"
        items:
          type: "string"
      recent_collaborators_id:
        type: "array"
        description: "array of user_ids of recent collaborators"
        items:
          type: "string"
      review_ids:
        type: "array"
        description: "array of `Review`ids"
        items:
          type: "string"
      state:
        type: "string"
        description: "State of `User`location"
      university_ids:
        type: "array"
        description: "array of universities the `User`is associated with"
        items:
          type: "string"
      updated_date:
        type: "string"
        format: "date-time"
        description: "datetime stamp of `User`update"
      user_id:
        type: "string"
        description: "Unique identifier representing a specific user"
      user_skill_ids:
        type: "array"
        description: "array of `UserSkill`ids"
        items:
          type: "string"
...